♻️ refactor(#64): Ticket -> Reservation 변경
