+++ sycls-user-api/src/main/java/com/b1/reservation/ReservationService.java
@@ -1,13 +1,13 @@
-package com.b1.ticket;
+package com.b1.reservation;
 
 import com.b1.content.entity.Round;
 import com.b1.place.RoundHelper;
+import com.b1.reservation.dto.ReservationAddRequestDto;
+import com.b1.reservation.dto.ReservationAddResponseDto;
 import com.b1.seat.SeatGradeHelper;
 import com.b1.seat.SeatReservationLogHelper;
 import com.b1.seat.entity.SeatGrade;
 import com.b1.seat.entity.SeatReservationLog;
-import com.b1.ticket.dto.TicketAddRequestDto;
-import com.b1.ticket.dto.TicketAddResponseDto;
 import java.util.LinkedHashSet;
 import java.util.Set;
 import lombok.RequiredArgsConstructor;
@@ -15,22 +15,21 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-@Slf4j(topic = "Ticket Service")
+@Slf4j(topic = "Reservation Service")
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class TicketService {
+public class ReservationService {
 
     private final RoundHelper roundHelper;
     private final SeatGradeHelper seatGradeHelper;
     private final SeatReservationLogHelper seatReservationLogHelper;
 
     /**
-     * 티켓 등록
+     * 예매 등록
      *
-     * @return
      */
-    public TicketAddResponseDto addTicket(final TicketAddRequestDto requestDto) {
+    public ReservationAddResponseDto addReservation(final ReservationAddRequestDto requestDto) {
         Set<Long> seatGradeIds = requestDto.seatGradeIds();
         final Round selectedRound = roundHelper.getRound(requestDto.roundId());
 
@@ -44,7 +43,7 @@ public class TicketService {
             newSeatReservationLogs.add(SeatReservationLog.addSeatReservationLog(seatGrade));
         }
         seatReservationLogHelper.addAllSeatReservationLogs(newSeatReservationLogs);
-        return TicketAddResponseDto.of(selectedRound.getId(), seatGradesForSelectedRound);
+        return ReservationAddResponseDto.of(selectedRound.getId(), seatGradesForSelectedRound);
     }
