+++ sycls-user-api/src/main/java/com/b1/reservation/dto/ReservationAddRequestDto.java
@@ -1,11 +1,11 @@
-package com.b1.ticket.dto;
+package com.b1.reservation.dto;
 
 import com.b1.constant.DomainConstant;
 import jakarta.validation.constraints.NotNull;
 import jakarta.validation.constraints.Size;
 import java.util.Set;
 
-public record TicketAddRequestDto(
+public record ReservationAddRequestDto(
         @NotNull(message = "공얀 회차 정보를 입력해 주세요")
         Long roundId,
         @NotNull(message = "좌석 정보가 누락되었습니다.")
