--- sycls-user-api/src/main/java/com/b1/ticket/TicketHelper.java
@@ -1,12 +1,0 @@
-package com.b1.ticket;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-
-@Component
-@RequiredArgsConstructor
-public class TicketHelper {
-
-    private final TicketRepository ticketRepository;
-
-}
